// projects/bible/mobile/src/interfaces/users/renewToken.ts
// Generated by https://quicktype.io

export interface RenewTokenResponse {
  status: string;
  ok: boolean;
  code: number;
  token: string;
  data: IUser;
}

export interface IUser {
  id: number;
  uuid: string;
  username: string;
  email: string;
  role: string;
  photo: string;
  googleId: null;
  passwordChangedAt?: null;
  passwordResetToken?: null;
  passwordResetExpires?: null;
  active: boolean;
  createdAt?: string;
  updatedAt?: string;
}